+++ www/enums/DeviceOrientations.js
@@ -1,0 +1,5 @@
+/**
+ * DeviceOrientations
+ */
+exports.PORTRAIT = 1;
+exports.LANDSCAPE = 1;